requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
Flask-Swagger-UI==3.36.0
openai==0.10.1
python-dotenv==0.19.2
